Mirror PDP: single Suspense around collection header
Mirror the pattern from ProductDetailPage so the handle-based collection
data resolves once at a single Suspense boundary, with searchParams kept
as a derived promise that's awaited at leaves only:

- Inline CollectionStructuredData and the title block into a single
  CollectionHeader async server component. One outer <Suspense> covers
  the structured data + title together. On cache hit, both render in a
  single tick instead of double-flashing through two separate Suspense
  fallbacks.
- CollectionHeaderFallback now matches the live h1 + description
  dimensions exactly: h-9/h-10/h-12 placeholder for the h1 (text-3xl ⇒
  line-height 36px, sm:text-4xl ⇒ 40px, md:text-5xl ⇒ 48px) plus an
  h-6 mt-1 description placeholder. No more 4px row growth on hydration,
  and no toolbar shift when a collection has a description.
- Inline the toolbar + grid render directly into CollectionDetailPage,
  removing the now-redundant CollectionResultsSection wrapper. Filter
  count badge moves with it. Drops components/collections/results.tsx
  and components/collections/structured-data.tsx (now unused).

Caching note: getCollection is "use cache: remote", so the cache lookup
still incurs a network roundtrip on first navigation to a handle. To
eliminate the brief shell flash entirely on subsequent navigations,
generateStaticParams could prerender known handles — out of scope here.
